#819 moving new record button so it shows whilst the onboarding section is visible
